backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.30
alembic>=1.13.2
psycopg[binary]>=3.2.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.2.0
numpy>=1.26.0
pillow>=10.4.0
tensorflow>=2.16.0
keras>=3.0.0
httpx>=0.27.0

# ML Training Dependencies
tensorflow-datasets>=4.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0.0
